Added Gift tracking feature
Gift responses are now only provided in the tooltip if you have already gifted the NPC the specified gift. This can be disabled/enabled by changing playerKnowsAllGifts in /Mods/RelationshipTooltips/config.json.

Sadly this is not retroactive. But it will continue to add keep track of gifts given from this update.