Fixed Tooltip Spacing, Added naming for pets of other players, Release
Tooltip spacing has been reworked, padding is now dictated by TooltipContent, which wraps and does text layout. It's nested within Tooltip's frame.

Pets of other players will now be named in MP.

Got rid of text scaling on dialogue font - was not the real issue - was the fault of padding.